rename(feature/ujin/#53): 댓글 컴포넌트 폴더 생성 및 경로 수정
