--- instagram/src/components/CommentMenu.jsx
@@ -1,20 +1,0 @@
-import { CardMenu, Interactions } from "../styles/CommentMenu_style";
-import { ReactComponent as Likes } from "../assets/likes.svg";
-import { ReactComponent as Comments } from "../assets/comments.svg";
-import { ReactComponent as DM } from "../assets/dm.svg";
-import { ReactComponent as Bookmark } from "../assets/bookmark.svg";
-
-function CommentMenu() {
-  return (
-    <CardMenu>
-      <Interactions>
-        <Likes className="icon" />
-        <Comments className="icon" />
-        <DM className="icon" />
-      </Interactions>
-      <Bookmark className="icon" />
-    </CardMenu>
-  );
-}
-
-export default CommentMenu;
