+++ instagram/src/components/LoginBody.jsx
@@ -1,29 +1,29 @@
-import React from "react";
-import Input from "./Input";
-import { Link } from "react-router-dom";
-import {
-  LoginBox,
-  LoginContainer,
-  ImgLogo,
-  LoginBtn,
-} from "../styles/LoginPage_style";
+// import React from "react";
+// import Input from "./Input";
+// import { Link } from "react-router-dom";
+// import {
+//   LoginBox,
+//   LoginContainer,
+//   ImgLogo,
+//   LoginBtn,
+// } from "../styles/LoginPage_style";
 
-function LoginBody() {
+// function LoginBody() {
   // function handleSubmit(e) {
   //   e.preventDefault();
   //   console.log("click");
   // }
-  return (
-    <LoginBox>
-      <LoginContainer>
-        <ImgLogo alt="instagram logo" src="img/instagramLogo.png" />
-        <Input />
-        <Link to="/main">
-          <LoginBtn>로그인</LoginBtn>
-        </Link>
-      </LoginContainer>
-    </LoginBox>
-  );
-}
+//   return (
+//     <LoginBox>
+//       <LoginContainer>
+//         <ImgLogo alt="instagram logo" src="img/instagramLogo.png" />
+//         <Input />
+//         <Link to="/main">
+//           <LoginBtn>로그인</LoginBtn>
+//         </Link>
+//       </LoginContainer>
+//     </LoginBox>
+//   );
+// }
 
-export default LoginBody;
+// export default LoginBody;
